MainApp Refactor Just some minor refactoring of the android MainApp class
